+++ modules/mdlogo.py
@@ -1,0 +1,18 @@
+import modules.constants as const
+from modules.utils import eprint
+
+def mobydump_logo() -> None:
+        """ Prints the MobyDump logo """
+
+        eprint('\n', wrap=False)
+        eprint(r' /~~~~~~~~~~~~~)', wrap=False)
+        eprint(r'(                )', wrap=False)
+        eprint(r'\        (o)      )', wrap=False)
+        eprint(r' \______/          )', wrap=False)
+        eprint(r'     \____          )', wrap=False)
+        eprint(r'    ___ __\       )', wrap=False)
+        eprint(r'   /_/\_\____    )  ', wrap=False)
+        eprint(r'   __       /  /', wrap=False)
+        eprint(r'   \ \_____/  /', wrap=False)
+        eprint(r'   / ________/   ', wrap=False)
+        eprint(f'  /_/ MOBYDUMP {const.__version__}\n', wrap=False)